This compiles. It should also work...
