+++ include/ReportIrradiation.hh
@@ -8,21 +8,32 @@
 #include <Report.hh>
 #include "rootweb.hh"
 #include "SummaryTable.h"
+#include "VizardTools.hh"
 #include "AnalyzerVisitors/IrradiationPower.h"
+#include <TH2D.h>
+#include "AnalyzerTools.hh"
+#include "SimParms.h"
 
 class DetectorModule;
 class Tracker;
 
-class ReportIrradiation : public Report {
+class ReportIrradiation : public Report, private insur::AnalyzerTools {
 private:
   IrradiationPowerVisitor irradiation_;
-  std::map<std::string, SummaryTable> sensorsIrradiationPowerSummary_;
-  std::map<std::string, SummaryTable> sensorsIrradiationSummary_;
-  void computeIrradiationPowerConsumption(const Tracker&);
+  std::map<std::string, SummaryTable> powerSummaries;
+  std::map<std::string, SummaryTable> irradiationSummaries;
+  void dumpRadiationTableSummary(RootWPage& myPage, std::map<std::string, SummaryTable>& radiationSummaries, const std::string& title, std::string units);
+  std::string createSensorsIrradiationCsv();
+  void computeIrradiationPowerConsumption();
   void preparePowerHistograms();
-
+  TH2D sensorsIrradiationPowerMap;
+  TH2D totalPowerConsumptionMap;
+  Tracker& tracker;
+  SimParms& simParms_;
 public:
-  void analyze(const Tracker&);
+  // TODO: avoid this :-)
+  ReportIrradiation(Tracker& t, SimParms& sp) : tracker(t), simParms_(sp) {};
+  void analyze();
   void visualizeTo(RootWSite&);
 };
