+++ include/Vizard.h
@@ -53,6 +53,7 @@
 #include <Palette.h>
 
 #include <PlotDrawer.h>
+#include "VizardTools.hh"
 
 namespace material {
   class WeightDistributionGrid;
@@ -71,23 +72,6 @@ namespace insur {
   static const std::string graph_wrong = "File stream reported error state: neighbour graph not written to file.";
   static const std::string exc_badalloc_graph = "Error: caught bad_alloc exception in Vizard::writeNeighbourGraph(). ";
   static const std::string graph_nowrite = "Neighbour graph was not written to file.";
-
-  // Some strings for the html formatting
-  static const std::string subStart   = "<sub>";      // These only should be needed
-  static const std::string subEnd     = "</sub>";
-  static const std::string superStart = "<sup>";
-  static const std::string superEnd   = "</sup>";
-  static const std::string smallStart = "<small>";
-  static const std::string smallEnd   = "</small>";
-  static const std::string emphStart  = "<b>";
-  static const std::string emphEnd    = "</b>";
-  static const std::string muLetter   = "&mu;";
-  static const std::string sigmaLetter= "&sigma;";
-  static const std::string etaLetter  = "&eta;";
-  static const std::string phiLetter  = "&phi;";
-  static const std::string thetaLetter= "&theta;";
-  static const std::string deltaLetter= "&delta;";
-  static const std::string tauLetter  = "&tau;";
 
   //clearStart="<tt>";
   //clearEnd="</tt>";
