+++ include/VizardTools.hh
@@ -1,0 +1,21 @@
+#ifndef VIZARDTOOLS_HH
+#define VIZARDTOOLS_HH
+
+// Some strings for the html formatting
+static const std::string subStart   = "<sub>";      // These only should be needed
+static const std::string subEnd     = "</sub>";
+static const std::string superStart = "<sup>";
+static const std::string superEnd   = "</sup>";
+static const std::string smallStart = "<small>";
+static const std::string smallEnd   = "</small>";
+static const std::string emphStart  = "<b>";
+static const std::string emphEnd    = "</b>";
+static const std::string muLetter   = "&mu;";
+static const std::string sigmaLetter= "&sigma;";
+static const std::string etaLetter  = "&eta;";
+static const std::string phiLetter  = "&phi;";
+static const std::string thetaLetter= "&theta;";
+static const std::string deltaLetter= "&delta;";
+static const std::string tauLetter  = "&tau;";
+
+#endif
